Move App inside Popup
