+++ source/Popup/index.tsx
@@ -3,7 +3,7 @@ import { ethers } from 'ethers';
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 import AggregatorClient from '../AggregatorClient';
-import App from '../App';
+import App from './App';
 import { AGGREGATOR_URL, CHAIN_ID, CHAIN_RPC_URL } from '../env';
 
 import Popup from './Popup';
